[Server] changed robot qos to floats

+++ trunk/Server/Controller/protocol.h
@@ -71,7 +71,7 @@ struct robotUpdate{
     int y;
     int listSize;
     int* objects;
-    int* qualities;
+    float* qualities;
     int* xs;
     int* ys;
     ~robotUpdate(){
